requirements.txt
pytorch>=1.0.0
torchvision>=0.2.0
numpy>=1.0.0
matplotlib>=3.0.0
tabulate>=0.8.7